a tool to list comments
